fixed some suggestions
