+++ app/models/fhir/types.py
@@ -1,8 +1,24 @@
+from dataclasses import dataclass
 from typing import Literal
 from enum import Enum
 from pydantic import BaseModel
 
 HttpValidVerbs = Literal["GET", "POST", "PATCH", "POST", "PUT", "HEAD", "DELETE"]
+
+ERROR_SEVERITIES = {"error", "fatal"}
+
+@dataclass
+class BundleError:
+    # Index in the bundle for this error
+    entry: int
+    # Status code of the error (http status code)
+    status: int
+    # Severity of the error
+    severity: str
+    # Error code
+    code: str
+    # Any diagnostics
+    diagnostics: str|None
 
 
 class BundleRequestParams(BaseModel):
